+++ packages/frontend/src/views/main/group/providers/ProvidersListView.vue
@@ -164,7 +164,9 @@ export default class ProvidersListView extends Vue {
 
   async deleteProvider(id: number) {
     if (self.confirm("Are you sure you want to delete the provider?")) {
-      await this.providerContext.actions.deleteProvider(id);
+      if (self.confirm("All children entities will be deleted!")) {
+        await this.providerContext.actions.deleteProvider(id);
+      }
     }
   }
 }
